refactor: rename carousel provider to carousel context
Replaced references to `provider` with `context` across carousel components for consistency with the existing naming conventions. Updated all imports, exports, and usages to reflect the change.
