+++ apps/nextjs/src/components/ui/carousel/carouse-right-chevron.tsx
@@ -2,7 +2,7 @@ import { useCallback } from 'react';
 import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
 import { IconButton } from '@/components/ui/icon-button';
-import { useCarouselContext } from './provider';
+import { useCarouselContext } from './context';
 
 export function CarouselRightChevron() {
   const { count, currentIndex, setCurrentIndex } = useCarouselContext();
