+++ apps/nextjs/src/components/ui/carousel/carousel-body.tsx
@@ -1,7 +1,7 @@
 import type { PropsWithChildren, ReactElement } from 'react';
 import { Children, cloneElement, isValidElement, useEffect } from 'react';
 import { Flex } from '@/components/ui/flex';
-import { useCarouselContext } from './provider';
+import { useCarouselContext } from './context';
 
 export function CarouselBody({ children }: PropsWithChildren) {
   const { setCount } = useCarouselContext();
