+++ apps/nextjs/src/components/ui/carousel/carousel-thumbnail-item.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo } from 'react';
 import { AspectRatio } from '@/components/ui/aspect-ratio';
 import { WrapItem } from '@/components/ui/wrap';
 import { useClickableHoverStyle } from '@/hooks/styles/use-clickable-hover-style';
-import { useCarouselContext } from './provider';
+import { useCarouselContext } from './context';
 
 type Props = {
   index: number;
